+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 v0.0.5
 2023-03-01
 * Added primary flag to wallet content type
+2023-03-02
+* Added wallet attach and detach API